+++ tests/main/clients/shorter_results/expected_client/fragments.py
@@ -1,8 +1,17 @@
+from typing import List
+
+from pydantic import Field
+
 from .base_model import BaseModel
 
 
-class AnimalFragment(BaseModel):
+class ListAnimalsFragment(BaseModel):
+    list_animals: List["ListAnimalsFragmentListAnimals"] = Field(alias="listAnimals")
+
+
+class ListAnimalsFragmentListAnimals(BaseModel):
     name: str
 
 
-AnimalFragment.update_forward_refs()
+ListAnimalsFragment.update_forward_refs()
+ListAnimalsFragmentListAnimals.update_forward_refs()