+++ tests/main/clients/shorter_results/expected_client/get_animal_fragment.py
@@ -1,8 +1,0 @@
-from .fragments import AnimalFragment
-
-
-class GetAnimalFragment(AnimalFragment):
-    pass
-
-
-GetAnimalFragment.update_forward_refs()